fix(relay): force dynamic rendering on setup-gated routes
`/`, `(admin)/layout`, and `/setup` decide their redirects from
`settingsStore().isSetupComplete()`, which reads `settings.json` from
disk. Next.js 15 was free to prerender these at `npm run build` time,
when the container has no `settings.json` yet — so the
"redirect to /setup" outcome got baked in as a static response. After
deployment the user could complete setup successfully (route handlers
run live, see the populated store, and reject re-setup with 409), but
every page-level navigation still served the stale static redirect,
trapping them on /setup.

Fix:

- Mark `/` and `(admin)/layout` as `dynamic = "force-dynamic"` so the
  setup gate is re-evaluated per request.
- Split `/setup` into a server guard + client form, and mark it
  dynamic too. The guard redirects to `/login` when setup is already
  complete, so the page can never trap a fully-initialized relay.
